Moved much of the field import logic to the fieldfactory and related schematic field models
