+++ models/Schematic_MatrixFieldModel.php
@@ -18,6 +18,9 @@ class Schematic_MatrixFieldModel extends Schematic_FieldModel
         return craft()->matrix;
     }
 
+    //==============================================================================================================
+    //================================================  EXPORT  ====================================================
+    //==============================================================================================================
 
     /**
      * @param FieldModel $field
@@ -59,5 +62,87 @@ class Schematic_MatrixFieldModel extends Schematic_FieldModel
         }
 
         return $blockTypeDefinitions;
+    }
+
+    //==============================================================================================================
+    //================================================  IMPORT  ====================================================
+    //==============================================================================================================
+
+    /**
+     * @param array $fieldDefinition
+     * @param FieldModel $field
+     * @param string $fieldHandle
+     * @param FieldGroupModel|null $group
+     */
+    public function populate(array $fieldDefinition, FieldModel $field, $fieldHandle, FieldGroupModel $group = null)
+    {
+        parent::populate($fieldDefinition, $field, $fieldHandle, $group);
+
+        /** @var MatrixSettingsModel $settingsModel */
+        $settingsModel = $field->getFieldType()->getSettings();
+        $settingsModel->setAttributes($fieldDefinition['settings']);
+        $settingsModel->setBlockTypes($this->getBlockTypes($fieldDefinition, $field));
+        $field->settings = $settingsModel;
+    }
+
+    /**
+     * Get blocktypes.
+     *
+     * @param array $fieldDefinition
+     * @param FieldModel $field
+     *
+     * @return mixed
+     */
+    protected function getBlockTypes(array $fieldDefinition, FieldModel $field)
+    {
+        $blockTypes = craft()->matrix->getBlockTypesByFieldId($field->id, 'handle');
+
+        foreach ($fieldDefinition['blockTypes'] as $blockTypeHandle => $blockTypeDef) {
+            $blockType = array_key_exists($blockTypeHandle, $blockTypes)
+                ? $blockTypes[$blockTypeHandle]
+                : new MatrixBlockTypeModel();
+
+            $this->populateBlockType($field, $blockType, $blockTypeDef, $blockTypeHandle);
+
+            $blockTypes[$blockTypeHandle] = $blockType;
+        }
+
+        return $blockTypes;
+    }
+
+    /**
+     * Populate blocktype.
+     *
+     * @param FieldModel $field
+     * @param MatrixBlockTypeModel $blockType
+     * @param array $blockTypeDef
+     * @param string $blockTypeHandle
+     */
+    private function populateBlockType(FieldModel $field, MatrixBlockTypeModel $blockType, array $blockTypeDef, $blockTypeHandle)
+    {
+        $fieldFactory = $this->getFieldFactory();
+
+        $blockType->fieldId = $field->id;
+        $blockType->name = $blockTypeDef['name'];
+        $blockType->handle = $blockTypeHandle;
+
+        $blockTypeFields = array();
+        foreach ($blockType->getFields() as $blockTypeField) {
+            $blockTypeFields[$blockTypeField->handle] = $blockTypeField;
+        }
+
+        $newBlockTypeFields = array();
+
+        foreach ($blockTypeDef['fields'] as $blockTypeFieldHandle => $blockTypeFieldDef) {
+            $blockTypeField = array_key_exists($blockTypeFieldHandle, $blockTypeFields)
+                ? $blockTypeFields[$blockTypeFieldHandle]
+                : new FieldModel();
+
+            $fieldFactory->populate($blockTypeFieldDef, $blockTypeField, $blockTypeFieldHandle);
+
+            $newBlockTypeFields[] = $blockTypeField;
+        }
+
+        $blockType->setFields($newBlockTypeFields);
     }
 }
